fix: do not modify the upstream-provided Docker systemd unit file
It's bad practice to modify files which are provided by a package, hence
rather provide a complete override unit based in the original one.

It's still questionable if this should be done at all just to
add directives relevant to TLS instead of using a Docker daemon
configuration file (`etc/docker/daemon.json`) but for now we stick with
this approach.